backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.1
sentence-transformers>=3.0.1
faiss-cpu>=1.9.0
numpy>=2.0.0
pandas>=2.2.2
groq>=0.9.0
edge-tts>=6.1.9
boto3>=1.34.0
python-multipart>=0.0.9
httpx>=0.27.0
pdfplumber>=0.11.0
PyMuPDF>=1.24.0
rank-bm25>=0.2.2
